+++ sph/include/sph/cuda/cuda_grad_p.cu
@@ -1,7 +1,7 @@
 #include <algorithm>
 
 #include "sph.cuh"
-#include "particles_data.hpp"
+#include "sph/particles_data.hpp"
 #include "cuda_utils.cuh"
 #include "sph/kernel/momentum_energy_kern.hpp"
