+++ sph/include/sph/propagator.hpp
@@ -34,20 +34,13 @@
 
 #include <iostream>
 
-#include "timer.hpp"
-#include "particles_data.hpp"
-
 #include "cstone/domain/domain.hpp"
 #include "ryoanji/cpu/treewalk.hpp"
 #include "ryoanji/cpu/upsweep.hpp"
+#include "sph/sph.hpp"
 
-#include "sph/av_switches.hpp"
-#include "sph/density_ve.hpp"
-#include "sph/divv_curlv.hpp"
-#include "sph/momentum_energy_ve.hpp"
-#include "sph/iad_ve.hpp"
-#include "sph/rho_zero.hpp"
-#include "sph/timestep.hpp"
+#include "util/timer.hpp"
+#include "particles_data.hpp"
 
 using namespace sphexa;
 using namespace sphexa::sph;
