+++ src/noh/noh_data_generator.hpp
@@ -4,7 +4,7 @@
 #include <cmath>
 #include <vector>
 
-#include "particles_data.hpp"
+#include "sph/particles_data.hpp"
 
 namespace sphexa
 {
